% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{crossover}
\alias{crossover}
\title{Genetic Operatior: Chromosome Crossover and Mutation}
\usage{
crossover(chr1,chr2)
}
\arguments{
\item{chr1, chr2}{a numeric vectors represents individual parent chromosome.}
}
\value{
A matrix with each column representing the
offsprings from a process of crossover and mutation.
}
\description{
Make 2 individual parent chromosomes crossover and
mutate when breeding offsprings.
}
\details{
This function makes two individual parent chromosomes
perfrom crossover and mutation when breeding next generation.
Note that the crossover is simply one-point crossover
and the mutation is based on \code{GA::mutation()}.
}
\examples{
ind1<-initialization(10)[,1]
ind2<-initialization(10)[,2]
crossover(ind1,ind2)
}
