% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/fitness.R
\name{fitness}
\alias{fitness}
\title{Calculate Fitness based on Fitness Function}
\usage{
fitness(pop, dat, fitnessFunction, model)

fitness(pop, dat, fitnessFunction, model)
}
\arguments{
\item{pop}{boleans matrix determined by \code{GA::initialization()}}

\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}

\item{fitnessFunction}{fitness function that takes in an lm or glm model and
returns a numerical fitness of that model. Users can choose AIC or BIC or
even define by themselves. If the argument is missing, the default is AIC.}

\item{model}{the linear model that user wants to use to fit in the data,
can be either \code{lm} or \code{glm}.}

\item{pop}{boleans matrix determined by \code{GA::initialization()}}

\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}

\item{fitnessFunction}{fitness function that takes in an lm or glm model and
returns a numerical fitness of that model. Users can choose AIC or BIC or
even define by themselves. If the argument is missing, the default is AIC.}

\item{model}{the linear model that user wants to use to fit in the data,
can be either \code{lm} or \code{glm}.}
}
\value{
Returns a matrix containing one row with \code{ncol(pop)}
observations of the fitness scores of each chromosomes.

Returns a matrix containing one row with \code{ncol(pop)} 
observations of the fitness scores of each chromosomes.
}
\description{
Calculate the fitness value for each individuals in the
generation based on the passed in fitness function.

Calculate the fitness value for each individuals in the
generation based on the passed in fitness function.
}
\examples{
dat <- mtcars
pop <- initialization(ncol(dat) - 1)
model <- lm
fitnessFunction <- AIC
fitness(pop, dat, fitnessFunction, model)
dat <- mtcars
pop <- initialization(ncol(dat) - 1)
model <- lm
fitnessFunction <- AIC
fitness(pop, dat, fitnessFunction, model)
}
