% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{initialization}
\alias{initialization}
\title{Randomly Generate the Initial Generation}
\usage{
initialization(C)
}
\arguments{
\item{C}{chromosome length (number of predictor variables in the data)}
}
\value{
A Bolean Matrix with dimension C by 1.5*C where each column
representing a chromosome, in which T marks that the gene (variable)
as active and F as inactive.
}
\description{
Bernoulli sampling to form the initial generation of the
Genetic Algorithm. The size of the initial generation is set to be the
integer closest to P = 1.5*C.
}
\details{
This function produces initial generation given the chromosome length
for Genetic Algorithm. The row of the generated boolean matrix represents the
locus of a gene(variable), and the column of the matrix represents different
members of the first generation(different models). The number of first
generation is defined to be integer closest to 1.5C.
}
\examples{
initialization(10)
}
