% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{nextGeneration}
\alias{nextGeneration}
\title{Breed the Selected Parents Generated From \code{GA::selection()}}
\usage{
nextGeneration(pop,selectedParents)
}
\arguments{
\item{pop}{boleans matrix determined by \code{GA::initialization()}}

\item{selectedParents}{list returnd by \code{GA::selection()}
At the minimum, it needs to contain indices for the selected parents 1 and 2.}

\item{...}{additional arguments to pass to the model function. DID NOT IMPLEMENT}
}
\value{
Returns a bolean matrix to replace the origin pop matrix
}
\description{
Breed the Selected Parents Generated From \code{GA::selection()}
}
\details{
Breeding uses \code{GA::crossover()} for each pair of parents
The first child is chosen to replace that pair of parents
Children chromosomes replace all parent chromosomes in the population
}
\examples{
dat <- mtcars # use the built-in mtcars data
C <- dim(dat)[2] - 1 #Number of variables
pop <- initialization(C) #produce boleans matrix
selectedParents <- selection(pop, f = AIC, dat)
pop <- nextGeneration(pop, selectedParents)
}
