% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Uses a genetic algorithm for variable selection in either lm or glm models}
\usage{
select(mod,dat,f)
}
\arguments{
\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}

\item{f}{fitness function that takes in an lm or glm model and returns a
a numerical 'qualification' of that model. Users can choose AIC or BIC or
even define by themselves. If the f argument is missing, the default is AIC.}

\item{model}{the linear model that user wants to use to fit in the data,
can be either \code{lm} or \code{glm}.}

\item{...}{additional arguments to pass to the model function. DID NOT IMPLEMENT YET}
}
\value{
Returns a list with the fittest model and
a matrix of the population fitness across generations (useful for plotting)
}
\description{
Uses a genetic algorithm for variable selection in either lm or glm models
}
\details{
The algorithm:
(1) First initializes population,
For g generations; do:
(2) calculates fitness of models and selects parent pairs to breed
(3) breeds the parent pairs, choosing the first child
(4) replaces the parents with the children
}
\examples{
dat <- mtcars
mod <- lm
library('stats')
f <- function(fit,...){return(extractAIC(fit,...)[2])}
results <- select(mod, dat, f)
summary(results$fittest_model)
X <- matrix(rep(seq(numGens), nrow(results$fitness), numGens), nrow = n, ncol = numGens)
Y <- t(results$fitness)
plot(X, Y, xlab = 'generation', ylab = 'AIC', pch = 19, cex = 0.5)
lines(seq(numGens), apply(results$fitness, FUN = min, MARGIN = 2), lty = 1, col = 'green')
}
